requirements.txt
numpy
pandas
matplotlib
seaborn
kaggle
kagglehub
scikit-learn
mlflow
supabase
ipykernel
streamlit
fastapi
uvicorn
setuptools
pytest
python-dotenv
datasets